+++ deployment/ccip/changeset/v1_6/cs_ccip_home_test.go
@@ -7,11 +7,15 @@ import (
 	"time"
 
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+	"github.com/ethereum/go-ethereum/common"
 	chain_selectors "github.com/smartcontractkit/chain-selectors"
 	"github.com/stretchr/testify/assert"
 	"golang.org/x/exp/maps"
 
+	"github.com/smartcontractkit/chainlink-ccip/pluginconfig"
+	"github.com/smartcontractkit/chainlink-common/pkg/types/ccipocr3"
 	cldf_chain "github.com/smartcontractkit/chainlink-deployments-framework/chain"
+	"github.com/smartcontractkit/chainlink-deployments-framework/datastore"
 
 	"github.com/smartcontractkit/chainlink-ccip/chainconfig"
 	cciptypes "github.com/smartcontractkit/chainlink-ccip/pkg/types/ccipocr3"
@@ -548,4 +552,116 @@ func Test_UpdateChainConfigs(t *testing.T) {
 			assert.Equal(t, chainConfigAfter2.Config, otherChainConfig.Config)
 		})
 	}
+}
+
+func Test_SetCandidateAcceptsUSDCTokenPoolProxyFromDataStore(t *testing.T) {
+	t.Parallel()
+
+	tenv, _ := testhelpers.NewMemoryEnvironment(t,
+		testhelpers.WithNumOfChains(2),
+		testhelpers.WithNumOfNodes(4))
+	dest := remoteChainSelector(
+		tenv.Env.BlockChains.ListChainSelectors(cldf_chain.WithFamily(chain_selectors.FamilyEVM)),
+		tenv.HomeChainSel,
+	)
+	proxyAddress := common.HexToAddress("0x1000000000000000000000000000000000000001")
+
+	ds := datastore.NewMemoryDataStore()
+	require.NoError(t, ds.Addresses().Add(datastore.AddressRef{
+		ChainSelector: dest,
+		Address:       proxyAddress.Hex(),
+		Type:          datastore.ContractType(shared.USDCTokenPoolProxy),
+		Version:       &deployment.Version1_7_0,
+		Qualifier:     "proxy-only-in-datastore",
+	}))
+	tenv.Env.DataStore = ds.Seal()
+
+	_, err := commonchangeset.Apply(t, tenv.Env,
+		commonchangeset.Configure(
+			cldf.CreateLegacyChangeSet(v1_6.SetCandidateChangeset),
+			setCandidateExecConfig(tenv.HomeChainSel, tenv.FeedChainSel, dest, proxyAddress.Hex()),
+		),
+	)
+	require.NoError(t, err)
+}
+
+func Test_SetCandidateErrorsOnDuplicateUSDCTokenPoolProxyInDataStore(t *testing.T) {
+	t.Parallel()
+
+	tenv, _ := testhelpers.NewMemoryEnvironment(t,
+		testhelpers.WithNumOfChains(2),
+		testhelpers.WithNumOfNodes(4))
+	dest := remoteChainSelector(
+		tenv.Env.BlockChains.ListChainSelectors(cldf_chain.WithFamily(chain_selectors.FamilyEVM)),
+		tenv.HomeChainSel,
+	)
+
+	ds := datastore.NewMemoryDataStore()
+	for i, ref := range []struct {
+		address   string
+		qualifier string
+	}{
+		{address: "0x1000000000000000000000000000000000000001", qualifier: "proxy-a"},
+		{address: "0x2000000000000000000000000000000000000002", qualifier: "proxy-b"},
+	} {
+		require.NoError(t, ds.Addresses().Add(datastore.AddressRef{
+			ChainSelector: dest,
+			Address:       common.HexToAddress(ref.address).Hex(),
+			Type:          datastore.ContractType(shared.USDCTokenPoolProxy),
+			Version:       &deployment.Version1_7_0,
+			Qualifier:     ref.qualifier,
+		}), "add datastore ref %d", i)
+	}
+	tenv.Env.DataStore = ds.Seal()
+
+	_, err := commonchangeset.Apply(t, tenv.Env,
+		commonchangeset.Configure(
+			cldf.CreateLegacyChangeSet(v1_6.SetCandidateChangeset),
+			setCandidateExecConfig(tenv.HomeChainSel, tenv.FeedChainSel, dest, "0x1000000000000000000000000000000000000001"),
+		),
+	)
+	require.Error(t, err)
+	require.ErrorContains(t, err, "multiple datastore entries found for USDCTokenPoolProxy 1.7.0")
+}
+
+func setCandidateExecConfig(homeChainSel, feedChainSel, dest uint64, sourcePoolAddress string) v1_6.SetCandidateChangesetConfig {
+	return v1_6.SetCandidateChangesetConfig{
+		SetCandidateConfigBase: v1_6.SetCandidateConfigBase{
+			HomeChainSelector: homeChainSel,
+			FeedChainSelector: feedChainSel,
+		},
+		PluginInfo: []v1_6.SetCandidatePluginInfo{
+			{
+				PluginType: types.PluginTypeCCIPExec,
+				OCRConfigPerRemoteChainSelector: map[uint64]v1_6.CCIPOCRParams{
+					dest: v1_6.DeriveOCRParamsForExec(v1_6.SimulationTest, []pluginconfig.TokenDataObserverConfig{
+						{
+							Type:    pluginconfig.USDCCCTPHandlerType,
+							Version: "1.0",
+							USDCCCTPObserverConfig: &pluginconfig.USDCCCTPObserverConfig{
+								AttestationConfig: pluginconfig.AttestationConfig{
+									AttestationAPI: "http://example.com",
+								},
+								Tokens: map[ccipocr3.ChainSelector]pluginconfig.USDCCCTPTokenConfig{
+									ccipocr3.ChainSelector(dest): {
+										SourcePoolAddress:            sourcePoolAddress,
+										SourceMessageTransmitterAddr: common.HexToAddress("0x3000000000000000000000000000000000000003").Hex(),
+									},
+								},
+							},
+						},
+					}, nil),
+				},
+			},
+		},
+	}
+}
+
+func remoteChainSelector(selectors []uint64, exclude uint64) uint64 {
+	for _, selector := range selectors {
+		if selector != exclude {
+			return selector
+		}
+	}
+	return 0
 }
